+++ src/test/scala/uk/gov/hmrc/govukfrontend/views/components/govukSummaryListSpec.scala
@@ -17,11 +17,12 @@
 package uk.gov.hmrc.govukfrontend.views.components
 
 import play.twirl.api.HtmlFormat
-import uk.gov.hmrc.govukfrontend.views.TemplateUnitSpec
+import uk.gov.hmrc.govukfrontend.views.{JsoupHelpers, TemplateUnitSpec}
 import uk.gov.hmrc.govukfrontend.views.html.components._
+
 import scala.util.Try
 
-class govukSummaryListSpec extends TemplateUnitSpec[SummaryList]("govukSummaryList") {
+class govukSummaryListSpec extends TemplateUnitSpec[SummaryList]("govukSummaryList") with JsoupHelpers {
 
   /**
     * Calls the Twirl template with the given parameters and returns the resulting markup
@@ -38,7 +39,9 @@ class govukSummaryListSpec extends TemplateUnitSpec[SummaryList]("govukSummaryLi
       val rowWithNoAction = SummaryListRow(key = Key(HtmlContent("foo")), value = Value(HtmlContent("FOO")), actions = None)
       val rowWithAction = SummaryListRow(key = Key(HtmlContent("bar")), value = Value(HtmlContent("BAR")), actions = Some(Actions(items = List(ActionItem(content = HtmlContent("link"))))))
       val html = GovukSummaryList(SummaryList(rowWithNoAction :: rowWithAction :: Nil))
-      println(html)
+
+      html.select(".govuk-summary-list__row") should have size 2
+      html.select(".govuk-link") should have size 1
     }
   }
 }